@@ -3,23 +3,20 @@
 # Licensed under the MIT license. See LICENSE file in the project root for details.
 #
 
-cmake_minimum_required(VERSION 3.1)
+cmake_minimum_required(VERSION 3.8)
 project(raw_image_pipeline)
 
-set(CMAKE_CXX_STANDARD 14)
-set(CMAKE_CXX_FLAGS "-std=c++14 -Wall -Wextra -O3 -DNDEBUG")
+set(CMAKE_CXX_STANDARD 17)
+if(NOT CMAKE_CONFIGURATION_TYPES AND NOT CMAKE_BUILD_TYPE)
+  set(CMAKE_BUILD_TYPE Release)
+endif()
+if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  add_compile_options(-Wall -Wextra -Wpedantic)
+endif()
 
-set(PACKAGE_DEPENDENCIES
-  eigen_conversions
-  raw_image_pipeline_white_balance
-  pybind11_catkin
-)
-
-find_package(catkin REQUIRED
-    COMPONENTS
-    ${PACKAGE_DEPENDENCIES}
-)
-# YAML-cpp
+find_package(ament_cmake REQUIRED)
+find_package(raw_image_pipeline_white_balance REQUIRED)
+# find_package(pybind11_catkin REQUIRED)
 find_package(yaml-cpp REQUIRED)
 
 find_package(Eigen3 REQUIRED)
@@ -52,22 +49,28 @@ else()
   set(CUDA_INCLUDE_DIRS "")
 endif()
 
-catkin_package(
-  INCLUDE_DIRS include
-  LIBRARIES ${PROJECT_NAME}
-  CATKIN_DEPENDS
-    ${PACKAGE_DEPENDENCIES}
-  DEPENDS
-    OpenCV
-    YAML_CPP
-)
-
 include_directories(
   include
-  ${catkin_INCLUDE_DIRS}
-  ${OpenCV_INCLUDE_DIRS}
-  ${CUDA_INCLUDE_DIRS}
+  SYSTEM ${CUDA_INCLUDE_DIRS}
+  SYSTEM ${EIGEN3_INCLUDE_DIR}
+  SYSTEM ${OpenCV_INCLUDE_DIRS}
 )
+
+# catkin_package(
+#   INCLUDE_DIRS include
+#   LIBRARIES ${PROJECT_NAME}
+#   CATKIN_DEPENDS
+#     ${PACKAGE_DEPENDENCIES}
+#   DEPENDS
+#     OpenCV
+#     YAML_CPP
+# )
+# 
+# include_directories(
+#   include
+#   ${catkin_INCLUDE_DIRS}
+#   ${OpenCV_INCLUDE_DIRS}
+# )
 
 add_library(${PROJECT_NAME} SHARED
   src/raw_image_pipeline/modules/color_calibration.cpp
@@ -80,25 +83,32 @@ add_library(${PROJECT_NAME} SHARED
   src/raw_image_pipeline/modules/white_balance.cpp
   src/raw_image_pipeline/raw_image_pipeline.cpp
 )
-
+ament_target_dependencies(${PROJECT_NAME}
+  Eigen3
+  OpenCV
+  raw_image_pipeline_white_balance
+  yaml-cpp
+)
 target_link_libraries(${PROJECT_NAME}
-    ${catkin_LIBRARIES}
-    ${CUDA_npp_LIBRARIES}
-    ${OpenCV_LIBS}
-    ${YAML_CPP_LIBRARIES}
+  ${CUDA_npp_LIBRARIES}
 )
 
 # Mark executables and/or libraries for installation
 install(
   TARGETS ${PROJECT_NAME}
-  ARCHIVE DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  LIBRARY DESTINATION ${CATKIN_PACKAGE_LIB_DESTINATION}
-  RUNTIME DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin
 )
 
-# Mark cpp header files for installation
-install(
-  DIRECTORY include/${PROJECT_NAME}/
-  DESTINATION ${CATKIN_PACKAGE_INCLUDE_DESTINATION}
+## Mark cpp header files for installation
+install(DIRECTORY include/
+  DESTINATION include
   FILES_MATCHING PATTERN "*.hpp"
 )
+
+ament_export_include_directories(include)
+ament_export_libraries(${PROJECT_NAME})
+ament_export_dependencies(Eigen3 OpenCV raw_image_pipeline_white_balance yaml-cpp)
+
+ament_package()
